Add a callback example
It's a common idiom for C callback interfaces to accept user context data in the form of a raw void pointer. This example demonstrates passing a Swift closure as that data. It is important to note that the capturing of data means that the Swift closure cannot be directly bridged. One unfortunate consequence is that passRetained will only take objects, so the closure must be wrapped in one